# backend cost pro/05_openapi.yaml
openapi: 3.1.0
info:
  title: API de Presupuestos de Obra
  version: 1.0.0
  description: |
    API derivada de las historias de usuario del DOC-02 y de las decisiones de
    diseno adoptadas en el DOC-03.

    Patrones transversales:
      - Confirmacion de impacto en dos fases (dryRun -> X-Confirmation-Token)
      - Reversibilidad con undoToken y endpoint generico POST /undo/{token}
      - Concurrencia optimista con ETag e If-Match
      - Validacion en dos capas: 422 para errores duros, arreglo `avisos` para blandos
      - Analitica calculada bajo demanda, con computedAt y ETag
  contact:
    name: Equipo de producto
servers:
  - url: https://api.example.com/v1
    description: Produccion
  - url: https://api.staging.example.com/v1
    description: Pruebas
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Proyectos]
      summary: Obtener proyecto
      description: HU-01
      responses:
        '200':
          description: Encabezado y estado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proyecto' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    patch:
      tags: [Proyectos]
      summary: Editar encabezado del proyecto
      description: HU-01. Requiere If-Match.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProyectoInput' }
      responses:
        '200':
          description: Proyecto actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proyecto' }
        '412': { $ref: '#/components/responses/ConflictoVersion' }
    delete:
      tags: [Proyectos]
      summary: Enviar proyecto a la papelera
      description: HU-03. Exige confirmacion escribiendo el nombre.
      parameters:
        - $ref: '#/components/parameters/ConfirmationToken'
      responses:
        '200':
          description: Enviado a papelera
          content:
            application/json:
              schema:
                type: object
                properties:
                  undoToken: { type: string }
        '428': { $ref: '#/components/responses/ConfirmacionRequerida' }
  /projects/{projectId}/duplicate:
    parameters:
      - $ref: '#/components/parameters/projectId'
    post:
      tags: [Proyectos]
      summary: Duplicar proyecto
      description: HU-03. Copia presupuesto, cantidades y parametros de AIU.
      responses:
        '201':
          description: Proyecto duplicado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proyecto' }
...
  /projects/{projectId}/apply-template:
    parameters:
      - $ref: '#/components/parameters/projectId'
    post:
      tags: [Proyectos]
      summary: Aplicar una plantilla al presupuesto
      description: HU-02. Operacion atomica. El modo es obligatorio si ya hay contenido.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId]
              properties:
                templateId: { type: string }
                modo:
                  type: string
                  enum: [REEMPLAZAR, AGREGAR]
      responses:
        '200':
          description: Plantilla aplicada
          content:
            application/json:
              schema:
                type: object
                properties:
                  actividadesCargadas: { type: integer }
                  actividadesOmitidas:
                    type: array
                    items:
                      type: object
                      properties:
                        apuId: { type: string }
                        motivo: { type: string }
                  totales: { $ref: '#/components/schemas/Totales' }
                  undoToken: { type: string }
        '409': { $ref: '#/components/responses/Conflicto' }
...
  /projects/{projectId}/aiu:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Presupuesto]
      summary: Configuracion de AIU vigente
      description: HU-16
      responses:
        '200':
          description: AIU
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Aiu' }
    put:
      tags: [Presupuesto]
      summary: Actualizar AIU, IVA y descuento
      description: |
        HU-16. Los tres porcentajes se aplican de forma independiente sobre el
        costo directo, nunca en cascada uno sobre otro.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Aiu' }
      responses:
        '200':
          description: AIU actualizado con totales recalculados
          content:
            application/json:
              schema:
                type: object
                properties:
                  aiu: { $ref: '#/components/schemas/Aiu' }
                  totales: { $ref: '#/components/schemas/Totales' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
      examples:
        default: { value: proj_9f2a }
...
    IfMatch:
      name: If-Match
      in: header
      description: ETag del recurso, para control de concurrencia optimista
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Clave generada por el cliente para evitar duplicados
      schema: { type: string }
    ConfirmationToken:
      name: X-Confirmation-Token
      in: header
      description: Token devuelto por el analisis de impacto previo
      schema: { type: string }
  responses:
...
    NoEncontrado:
      description: El recurso no existe o pertenece a otra organizacion
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflicto:
      description: El estado actual impide la operacion
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ConflictoVersion:
      description: El If-Match no coincide, otro usuario modifico el recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ConfirmacionRequerida:
      description: Falta el X-Confirmation-Token de una operacion con impacto
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Dinero:
      type: string
      description: Cadena decimal con dos decimales. Nunca coma flotante.
      pattern: '^-?\d+\.\d{2}$'
      examples: ['58696983.45']
...
    Error:
      type: object
      properties:
        tipo:
          type: string
          enum: [PETICION_INVALIDA, NO_AUTENTICADO, SIN_PERMISO, NO_ENCONTRADO,
                 CONFLICTO, EXPIRADO, CONFLICTO_VERSION, VALIDACION,
                 CONFIRMACION_REQUERIDA, CUOTA_AGOTADA, DEMASIADAS_PETICIONES,
                 SERVICIO_NO_DISPONIBLE]
        mensaje: { type: string, description: Apto para mostrar al usuario }
        traceId: { type: string }
        errores:
          type: array
          items:
            type: object
            properties:
              campo: { type: string }
              codigo: { type: string }
              esperado: { type: string }
    Totales:
      type: object
      properties:
        costoDirecto: { $ref: '#/components/schemas/Dinero' }
        aiu: { $ref: '#/components/schemas/Dinero' }
        iva: { $ref: '#/components/schemas/Dinero' }
        descuento: { $ref: '#/components/schemas/Dinero' }
        total: { $ref: '#/components/schemas/Dinero' }
    ProyectoInput:
      type: object
      required: [nombre, areaM2]
      properties:
        nombre: { type: string, maxLength: 120 }
        cliente: { type: string, maxLength: 120 }
        ubicacion: { type: string, maxLength: 160 }
        areaM2: { type: number, exclusiveMinimum: 0 }
        tipoObra:
          type: string
          enum: [OBRA_NUEVA, REMODELACION, ADECUACION, MANTENIMIENTO]
        fecha: { type: string, format: date }
        descuento: { type: number, minimum: 0 }
    Proyecto:
      allOf:
        - $ref: '#/components/schemas/ProyectoInput'
        - type: object
          properties:
            id: { type: string }
            estado:
              type: string
              enum: [BORRADOR, EN_REVISION, APROBADO, ARCHIVADO, PAPELERA]
            costoDirecto: { $ref: '#/components/schemas/Dinero' }
            total: { $ref: '#/components/schemas/Dinero' }
            creadoEn: { type: string, format: date-time }
            modificadoEn: { type: string, format: date-time }
            version: { type: integer }
...
    Aiu:
      type: object
      properties:
        pctAdministracion: { type: number, minimum: 0, maximum: 100 }
        pctImprevistos: { type: number, minimum: 0, maximum: 100 }
        pctUtilidad: { type: number, minimum: 0, maximum: 100 }
        ivaAplica: { type: boolean }
        pctIva: { type: number, minimum: 0, maximum: 100, default: 19 }
        baseIva:
          type: string
          enum: [TOTAL, AIU, UTILIDAD]
        descuento: { type: number, minimum: 0 }
...